Add files via upload
This Android application is designed to provide users with relevant transportation information through the use of affordable, small mobile devices that have limited computational resources, such as PDAs and smartphones. By utilizing common low-cost computing and telecommunications hardware, bus agencies and transportation authorities are able to support the infrastructure necessary for this application.

The Transport Tracking System, which relies on Wi-Fi technology, is capable of providing bus schedules, timing, and routing information to students of UMSIT who are in the vicinity of a Transport Tracking System Smart Bus Stop (SBS), which serves as a hotspot and information-gathering terminal node. PDA and smartphone users can freely access the Transport Tracking System service to obtain information about bus stops, including the arrival times of the next buses for a given destination or route, routing to various locations, and time-of-travel estimations. The hand-over of mobile device control is handled by a unique or specific protocol.

This information is periodically transmitted back to the bus stops, allowing users and students to obtain relevant information through a proposed method.